requirements.txt
aiohttp==3.7.4.post0
astunparse==1.6.3
async-timeout==3.0.1
asyncpg==0.23.0
attrs==21.2.0
braceexpand==0.1.7
chardet==4.0.0
click==8.0.1
discord.py==1.7.2
discord.py-stubs==1.7.1
humanize==3.7.0
idna==3.2
import-expression==1.1.4
jishaku==1.20.0.220
motor==2.4.0
multidict==5.1.0
mypy==0.812
mypy-extensions==0.4.3
pymongo==3.11.4
six==1.16.0
typed-ast==1.4.3
typing-extensions==3.10.0.0
yarl==1.6.3
